fix: factor out insert logic
