+++ pkg/repository/v1/payloadstore.go
@@ -444,76 +444,23 @@ func (p *payloadStoreRepositoryImpl) CopyOffloadedPayloadsIntoTempTable(ctx cont
 			retrieveOptsToKey[r] = k
 		}
 
-		tenantIds := make([]pgtype.UUID, 0, len(payloads))
-		ids := make([]int64, 0, len(payloads))
-		insertedAts := make([]pgtype.Timestamptz, 0, len(payloads))
-		externalIds := make([]pgtype.UUID, 0, len(payloads))
-		types := make([]string, 0, len(payloads))
-		locations := make([]string, 0, len(payloads))
-		externalLocationKeys := make([]string, 0, len(payloads))
+		payloadsToInsert := make([]sqlcv1.CutoverPayloadToInsert, 0, len(payloads))
 
 		for r, k := range retrieveOptsToKey {
-			// qq: do we need conflict resolution here? I think the `COPY` is probably fine
 			externalId := retrieveOptsToExternalId[r]
 
-			tenantIds = append(tenantIds, r.TenantId)
-			ids = append(ids, r.Id)
-			insertedAts = append(insertedAts, r.InsertedAt)
-			types = append(types, string(r.Type))
-			locations = append(locations, string(sqlcv1.V1PayloadLocationEXTERNAL))
-			externalLocationKeys = append(externalLocationKeys, string(k))
-			externalIds = append(externalIds, sqlchelpers.UUIDFromStr(externalId))
+			// todo: make sure location works for already-inserted payloads
+			payloadsToInsert = append(payloadsToInsert, sqlcv1.CutoverPayloadToInsert{
+				TenantID:            r.TenantId,
+				ID:                  r.Id,
+				InsertedAt:          r.InsertedAt,
+				ExternalID:          sqlchelpers.UUIDFromStr(externalId),
+				Type:                r.Type,
+				ExternalLocationKey: string(k),
+			})
 		}
 
-		row := tx.QueryRow(
-			ctx,
-			fmt.Sprintf(
-				// we unfortunately need to use `INSERT INTO` instead of `COPY` here
-				// because we can't have conflict resolution with `COPY`.
-				`
-				WITH inputs AS (
-					SELECT
-						UNNEST($1::UUID[]) AS tenant_id,
-						UNNEST($2::BIGINT[]) AS id,
-						UNNEST($3::TIMESTAMPTZ[]) AS inserted_at,
-						UNNEST($4::UUID[]) AS external_id,
-						UNNEST($5::TEXT[]) AS type,
-						UNNEST($6::TEXT[]) AS location,
-						UNNEST($7::TEXT[]) AS external_location_key
-				), inserts AS (
-					INSERT INTO %s (tenant_id, id, inserted_at, external_id, type, location, external_location_key, inline_content, updated_at)
-					SELECT
-						tenant_id,
-						id,
-						inserted_at,
-						external_id,
-						type::v1_payload_type,
-						location::v1_payload_location,
-						external_location_key,
-						NULL,
-						NOW()
-					FROM inputs
-					ORDER BY tenant_id, inserted_at, id, type
-					ON CONFLICT(tenant_id, id, inserted_at, type) DO NOTHING
-					RETURNING *
-				)
-
-				SELECT COUNT(*)
-				FROM inserts
-				`,
-				tableName,
-			),
-			tenantIds,
-			ids,
-			insertedAts,
-			externalIds,
-			types,
-			locations,
-			externalLocationKeys,
-		)
-
-		var copyCount int64
-		err = row.Scan(&copyCount)
+		_, err = sqlcv1.InsertCutOverPayloadsIntoTempTable(ctx, tx, tableName, payloadsToInsert)
 
 		if err != nil {
 			return fmt.Errorf("failed to copy offloaded payloads into temp table: %w", err)
